refactor: Merge `Video ads` into `Hide ads`
