--- extensions/music/src/main/java/app/morphe/extension/music/patches/HideVideoAdsPatch.java
@@ -1,17 +1,0 @@
-package app.morphe.extension.music.patches;
-
-import app.morphe.extension.music.settings.Settings;
-
-@SuppressWarnings("unused")
-public class HideVideoAdsPatch {
-
-    /**
-     * Injection point
-     */
-    public static boolean showVideoAds(boolean original) {
-        if (Settings.HIDE_VIDEO_ADS.get()) {
-            return false;
-        }
-        return original;
-    }
-}
