--- patches/src/main/kotlin/app/morphe/patches/music/ad/video/Fingerprints.kt
@@ -1,14 +1,0 @@
-package app.morphe.patches.music.ad.video
-
-import app.morphe.patcher.Fingerprint
-import app.morphe.patcher.OpcodesFilter
-import com.android.tools.smali.dexlib2.Opcode
-
-internal object ShowVideoAdsFingerprint : Fingerprint(
-    filters = OpcodesFilter.opcodesToFilters(
-        Opcode.MOVE_RESULT_OBJECT,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.IGET_OBJECT,
-    ),
-    strings = listOf("maybeRegenerateCpnAndStatsClient called unexpectedly, but no error.")
-)
