--- patches/src/main/kotlin/app/morphe/patches/music/ad/video/HideVideoAds.kt
@@ -1,38 +1,0 @@
-package app.morphe.patches.music.ad.video
-
-import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
-import app.morphe.patcher.patch.bytecodePatch
-import app.morphe.patches.music.misc.extension.sharedExtensionPatch
-import app.morphe.patches.music.misc.settings.PreferenceScreen
-import app.morphe.patches.music.misc.settings.settingsPatch
-import app.morphe.patches.music.shared.Constants.COMPATIBILITY_YOUTUBE_MUSIC
-import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
-
-private const val EXTENSION_CLASS = "Lapp/morphe/extension/music/patches/HideVideoAdsPatch;"
-
-@Suppress("unused")
-val hideVideoAdsPatch = bytecodePatch(
-    name = "Hide music video ads",
-    description = "Adds an option to hide ads that appear while listening to or streaming music videos, podcasts, or songs.",
-) {
-    dependsOn(
-        sharedExtensionPatch,
-        settingsPatch,
-    )
-
-    compatibleWith(COMPATIBILITY_YOUTUBE_MUSIC)
-
-    execute {
-        PreferenceScreen.ADS.addPreferences(
-            SwitchPreference("morphe_music_hide_video_ads"),
-        )
-
-        ShowVideoAdsFingerprint.instructionMatches[1].getMethodCalled().addInstructions(
-            0,
-            """
-                invoke-static { p1 }, $EXTENSION_CLASS->showVideoAds(Z)Z
-                move-result p1
-            """
-        )
-    }
-}
