--- patches/src/main/kotlin/app/morphe/patches/youtube/ad/video/Fingerprints.kt
@@ -1,18 +1,0 @@
-package app.morphe.patches.youtube.ad.video
-
-import app.morphe.patcher.Fingerprint
-
-internal object LoadVideoAdsFingerprint : Fingerprint(
-    strings = listOf(
-        "TriggerBundle doesn't have the required metadata specified by the trigger ",
-        "Ping migration no associated ping bindings for activated trigger: ",
-    )
-)
-
-internal object PlayerBytesAdLayoutFingerprint : Fingerprint(
-    returnType = "V",
-    parameters = listOf("L"),
-    strings = listOf(
-        "Bootstrapped layout construction resulted in non PlayerBytesLayout. PlayerAds count: ",
-    )
-)
